Example NIST 800-53 control ID to inspec mapper.
Per request from @gregelin, here is a quick one-off script to demo how
one can use Python to load the JSON from the example Ubuntu 16.04 LTS
Inspec profile to check conformance of a Ubuntu system with the DISA
STIG into a properly formed Python dictionary for further processing.

For now, it will dump to screen. This is preliminary and further reqs
can be built do discuss further processing and mapping of Inspec prof
content from Inspec to an OSCAL SAP and SAR model when we make more
progress.